requirements.txt
flask
google-genai
pillow
python-docx 
python-pptx 
pandas 
openpyxl 
PyPDF2